删除所有 DataDao 依赖
还未调试通过，暂不要下载

+++ Runtime/NPiculet.Base/EF/cms_points_log.cs
@@ -22,7 +22,7 @@ namespace NPiculet.Base.EF
         public string TargetId { get; set; }
         public string IP { get; set; }
         public string Tag { get; set; }
-        public Nullable<int> Point { get; set; }
+        public Nullable<decimal> Point { get; set; }
         public string Comment { get; set; }
         public Nullable<System.DateTime> CreateDate { get; set; }
         public string Creator { get; set; }
